+++ examples/src/nrf52_dk.rs
@@ -1,11 +1,10 @@
 #![cfg_attr(not(test), no_std)]
 #![cfg_attr(not(test), no_main)]
 
-extern crate panic_rtt;
-
 use core::convert::Infallible;
 use cortex_m_rt::entry;
 use embedded_time::{self as time, traits::*};
+use panic_rtt as _;
 
 pub mod nrf52 {
     pub use nrf52832_hal::{
@@ -37,8 +36,8 @@ impl SysClock {
 
 impl time::Clock for SysClock {
     type Rep = u64;
-    const PERIOD: time::Period = <time::Period>::new(1, 16_000_000);
     type ImplError = Infallible;
+    const PERIOD: time::Period = <time::Period>::new(1, 16_000_000);
 
     fn now(&self) -> Result<time::Instant<Self>, time::clock::Error<Self::ImplError>> {
         self.capture_task.tasks_trigger[0].write(|write| unsafe { write.bits(1) });
@@ -100,8 +99,8 @@ fn main() -> ! {
 
     // This moves these peripherals to prevent conflicting usage, however not the entire EGU0 is
     // used. A ref to EGU0 could be sent instead, although that provides no protection for the
-    // fields that are being used by Clock64.
-    let mut clock = SysClock::take(device.TIMER0, device.TIMER1, device.EGU0);
+    // fields that are being used by the clock.
+    let clock = SysClock::take(device.TIMER0, device.TIMER1, device.EGU0);
 
     let port0 = nrf52::gpio::p0::Parts::new(device.P0);
 
@@ -130,7 +129,7 @@ fn main() -> ! {
         &mut led2.degrade(),
         &mut led3.degrade(),
         &mut led4.degrade(),
-        &mut clock,
+        &clock,
     )
     .unwrap();
 
@@ -142,7 +141,7 @@ fn run<Led>(
     led2: &mut Led,
     led3: &mut Led,
     led4: &mut Led,
-    clock: &mut SysClock,
+    clock: &SysClock,
 ) -> Result<(), <Led as nrf52::OutputPin>::Error>
 where
     Led: nrf52::OutputPin,
@@ -152,12 +151,22 @@ where
         led2.set_high()?;
         led3.set_high()?;
         led4.set_low()?;
-        clock.new_timer(250_u32.milliseconds()).start().wait();
+        clock
+            .new_timer(250_u32.milliseconds())
+            .start()
+            .unwrap()
+            .wait()
+            .unwrap();
 
         led1.set_high()?;
         led2.set_low()?;
         led3.set_low()?;
         led4.set_high()?;
-        clock.new_timer(250_u32.milliseconds()).start().wait();
+        clock
+            .new_timer(250_u32.milliseconds())
+            .start()
+            .unwrap()
+            .wait()
+            .unwrap();
     }
 }