+++ src/frequency.rs
@@ -1,32 +1,14 @@
 //! Representations of frequency-based values
 
 pub(crate) mod units {
-    use crate::{Period, TimeInt};
+    use crate::{ConversionError, Period, TimeInt};
     use core::{convert, ops};
 
     /// A frequency unit type
     ///
     /// Convertible to/from [`Period`].
-    #[derive(Ord, PartialOrd, Eq, PartialEq, Debug)]
+    #[derive(Copy, Clone, Debug, Ord, PartialOrd, Eq, PartialEq)]
     pub struct Hertz<T: TimeInt = u32>(pub T);
-
-    impl<T: TimeInt> Hertz<T> {
-        /// ```rust
-        /// # use embedded_time::{Period, units::*};
-        /// assert_eq!(Hertz(1_000_u32).into_period(), <Period>::new(1, 1_000));
-        /// ```
-        pub fn into_period(self) -> Period<T> {
-            Period::from_frequency(self)
-        }
-
-        /// ```rust
-        /// # use embedded_time::{Period, units::*};
-        /// assert_eq!(Hertz::from_period(Period::new(1, 1_000)), Hertz(1_000_u32));
-        /// ```
-        pub fn from_period(period: Period<T>) -> Self {
-            period.to_frequency()
-        }
-    }
 
     impl<T: TimeInt, Rhs: TimeInt> ops::Mul<Rhs> for Hertz<T>
     where
@@ -55,6 +37,63 @@ pub(crate) mod units {
         /// ```
         fn div(self, rhs: Rhs) -> Self::Output {
             Self(self.0 / <T as convert::From<Rhs>>::from(rhs))
+        }
+    }
+
+    impl<T: TimeInt> convert::TryFrom<Period<T>> for Hertz<T> {
+        type Error = ConversionError;
+
+        /// Create a new `Frequency` from a fractional [`Period`] in seconds
+        ///
+        /// # Examples
+        ///
+        /// ```rust
+        /// # use embedded_time::{Period, units::*, ConversionError};
+        /// # use core::{convert::{TryFrom, TryInto}};
+        /// #
+        /// assert_eq!(Hertz::try_from(<Period>::new(1, 1_000)),
+        ///     Ok(Hertz(1_000_u32)));
+        ///
+        /// assert_eq!(Hertz::try_from(<Period>::new(0, 1_000)),
+        ///     Err(ConversionError::DivByZero));
+        ///
+        /// assert_eq!(<Period>::new(1, 1_000).try_into(),
+        ///     Ok(Hertz(1_000_u32)));
+        /// ```
+        ///
+        /// # Errors
+        ///
+        /// [`ConversionError::DivByZero`]
+        fn try_from(period: Period<T>) -> Result<Self, Self::Error> {
+            if !period.numerator().is_zero() {
+                Ok(Hertz(period.recip().to_integer()))
+            } else {
+                Err(ConversionError::DivByZero)
+            }
+        }
+    }
+
+    impl convert::From<u32> for Hertz {
+        /// ```rust
+        /// # use embedded_time::units::Hertz;
+        /// #
+        /// assert_eq!(Hertz::from(23), Hertz(23_u32));
+        /// assert_eq!(<Hertz as Into<u32>>::into(Hertz(23_u32)), 23_u32);
+        /// ```
+        fn from(value: u32) -> Self {
+            Self(value)
+        }
+    }
+
+    impl convert::From<Hertz> for u32 {
+        /// ```rust
+        /// # use embedded_time::units::Hertz;
+        /// #
+        /// assert_eq!(u32::from(Hertz(23)), 23_u32);
+        /// assert_eq!(<u32 as Into<Hertz>>::into(23), Hertz(23_u32));
+        /// ```
+        fn from(hertz: Hertz) -> Self {
+            hertz.0
         }
     }
 }