+++ src/time_int.rs
@@ -1,4 +1,4 @@
-use crate::Period;
+use crate::{ConversionError, Period};
 use core::{convert::TryFrom, convert::TryInto, fmt};
 
 /// The core inner-type trait for time-related types
@@ -8,6 +8,8 @@ pub trait TimeInt:
     + num::Bounded
     + num::traits::WrappingAdd
     + num::traits::WrappingSub
+    + num::CheckedAdd
+    + num::CheckedSub
     + num::CheckedMul
     + num::CheckedDiv
     + From<u32>
@@ -20,38 +22,46 @@ pub trait TimeInt:
     + fmt::Display
     + fmt::Debug
 {
-    /// A checked multiplication with a [`Period`]
+    /// Checked integer * [`Period`] = integer
+    ///
+    /// Returns truncated integer
     ///
     /// # Examples
     ///
     /// ```rust
     /// # use embedded_time::{Period, traits::*};
-    /// assert_eq!(8_u32.checked_mul_period(&Period::new(1,2)), Some(4_u32));
+    /// assert_eq!(8_u32.checked_mul_period(&<Period>::new(1,2)), Ok(4_u32));
     ///
     /// // the result is not rounded, but truncated
-    /// assert_eq!(8_u32.checked_mul_period(&Period::new(1,3)), Some(2_u32));
+    /// assert_eq!(8_u32.checked_mul_period(&<Period>::new(1,3)), Ok(2_u32));
     /// ```
-    fn checked_mul_period(&self, period: &Period) -> Option<Self> {
-        Some(<Self as num::CheckedDiv>::checked_div(
-            &<Self as num::CheckedMul>::checked_mul(&self, &(*period.numerator()).into())?,
+    fn checked_mul_period(&self, period: &Period) -> Result<Self, ConversionError> {
+        <Self as num::CheckedDiv>::checked_div(
+            &<Self as num::CheckedMul>::checked_mul(&self, &(*period.numerator()).into())
+                .ok_or(ConversionError::Overflow)?,
             &(*period.denominator()).into(),
-        )?)
+        )
+        .ok_or(ConversionError::DivByZero)
     }
 
-    /// A checked division with a [`Period`]
+    /// Checked integer / [`Period`] = integer
+    ///
+    /// Returns truncated integer
     ///
     /// # Examples
     ///
     /// ```rust
     /// # use embedded_time::{Period, traits::*};
-    /// assert_eq!(8_u32.checked_div_period(&Period::new(1,2)), Some(16_u32));
-    /// assert_eq!(8_u32.checked_div_period(&Period::new(3,2)), Some(5_u32));
+    /// assert_eq!(8_u32.checked_div_period(&<Period>::new(1,2)), Ok(16_u32));
+    /// assert_eq!(8_u32.checked_div_period(&<Period>::new(3,2)), Ok(5_u32));
     /// ```
-    fn checked_div_period(&self, period: &Period) -> Option<Self> {
-        Some(<Self as num::CheckedDiv>::checked_div(
-            &<Self as num::CheckedMul>::checked_mul(&self, &(*period.denominator()).into())?,
+    fn checked_div_period(&self, period: &Period) -> Result<Self, ConversionError> {
+        <Self as num::CheckedDiv>::checked_div(
+            &<Self as num::CheckedMul>::checked_mul(&self, &(*period.denominator()).into())
+                .ok_or(ConversionError::Overflow)?,
             &(*period.numerator()).into(),
-        )?)
+        )
+        .ok_or(ConversionError::DivByZero)
     }
 }